abstract out the separate chunks of saving (pt 2)
This was mostly already done in the previous commit just by ripping things out, so this one is focused on getting things hooked back together. Next commit is making sure it works. 2 commits of non-functional code.

Also removed an unnecessary import from notebook_frame.py
Also cleaned up the to-do list in the log file
