+++ log.md
@@ -16,15 +16,13 @@ If you see a blank line between bullets on any given day, it means that I rolled
 
 
 - is ``add_data_tab`` necessary anymore?
-- Breakdown ``save.py``
-    - probs use a sub_folder, as cumbersome as they are becoming, to create a class for ``grab data``, and a class for ``save``.
-    - it's getting way too long and it's kind of annoying me. there's so much for just grabbing the data, and so little for saving, but it's 275 lines long.
 - Reloading save data
-- clean up the code to remove some of the hackiness that arose from my brain turning into a stubborn husky
-- Update the ``state`` variable name for the checklist processing to instead be something more accurately referring to ``collection status``
-- Rename the user file / master save file to be ``master`` instead of ``time``
-- when collecting data, store also the tab type
-    - The val will be a list, index 0 is tab type and index 1 is tab contents
+- Improve some code in ``gather_tab_data.py`` (maybe even rename it)
+    - Update the ``state`` variable name for the checklist processing to instead be something more accurately referring to ``collection status``
+    - when collecting data, store also the tab type
+        - The val will be a list, index 0 is tab type and index 1 is tab contents
+- Improve some code in ``create_save_dir.py`` (maybe even rename it)
+    - Rename the user file / master save file to be ``master`` instead of ``time``, so the user knows what to look for
 - Figure out how to implement an autosave upon resets, and inform the user of the autosave
 - Proper errorhandling instead of ``raise NotImplementedError``
 - Textwrapping in the textboxes to be word
